Add two missing cases
