+++ packages/vscode-extension/src/extension/main.ts
@@ -27,6 +27,7 @@ import { registerFileSystemProvider } from "./file-system-provider";
 import { registerProgressReporter } from "./progress";
 import {
   deriveUserSettingsUri,
+  locateWorkspaceFolder,
   registerConfigLoader,
   watchPluginSettings,
 } from "./config-loader";
@@ -78,13 +79,13 @@ function registerOnDidOpenTextDocListener(
   const listener = async (document: vscode.TextDocument) => {
     // settle on the 1st workspace folder available
     // TODO @montymxb May 15th, 2025: Support configs across multiple workspace folders
-    const workspaceFolder = vscode.workspace.workspaceFolders?.[0]?.uri.fsPath;
+    const workspaceFolder = locateWorkspaceFolder(document.uri);
     if (!workspaceFolder) {
       return;
     }
 
     // check if we can create a .pliplugin folder
-    const plipluginPath = path.join(workspaceFolder, ".pliplugin");
+    const plipluginPath = path.join(workspaceFolder.fsPath, ".pliplugin");
     if (document.languageId !== "pli" || fs.existsSync(plipluginPath)) {
       // not a pli file or config already exists
       return;
